Java_Coding_Test_Practice 0925

+++ Code_test/src/baekjoon/implementations/bronze/b10824/Main.java
@@ -10,8 +10,8 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int N = Integer.parseInt(st.nextToken() + st.nextToken());
-        int M = Integer.parseInt(st.nextToken() + st.nextToken());
+        Long N = Long.parseLong(st.nextToken() + st.nextToken());
+        Long M = Long.parseLong(st.nextToken() + st.nextToken());
         System.out.println(N + M);
     }
